Move utilies into util package Move custom structures into structs package
