+++ frontend/structs/Coord.test.ts
