+++ frontend/structs/Coord.ts
