+++ frontend/structs/Freezable.test.ts
