+++ frontend/structs/Freezable.ts
