+++ frontend/structs/IFreezable.ts
