+++ frontend/tetris/TetrisState.test.ts
@@ -1,6 +1,6 @@
-import Coord from "./Coord";
-import Move from "./Move";
-import Position from "./Position";
+import Coord from "../structs/Coord";
+import Move from "../structs/Move";
+import Position from "../structs/Position";
 import { Shape } from "./Shape";
 import TetrisState from "./TetrisState";
